+++ resources/release-notes.md
@@ -1,6 +1,7 @@
 # Software
 
 - Improve handling of "croppable" data (inheriting from `mag.mixin.Croppable`) with utility methods `mustBeTimeFilter` and `convertToTimeSubscript`
+- Do not use "Primary" or "Secondary" in sensor event labels
 
 # GitHub